backend/requirements.txt
django>=5.2.7
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
scikit-learn>=1.3.0